// src/app/models/enum.product.actions.type.ts
export enum ProductActionsType {
  GET_ALL_PRODUCTS = "[Product] Get all products!",
  GET_SELECTED_PRODUCTS = "[Product] Get selected products!",
  GET_AVAILABLE_PRODUCTS = "[Product] Get available products!",
  NEW_PRODUCTS = "[Product] New products!",
  SEARCH_PRODUCTS = "[Product] search products!",
  SELECT_PRODUCTS = "[Product] Select product!",
  EDIT_PRODUCTS = "[Product] Edit product!",
  DELETE_PRODUCTS = "[Product] Delete product!"
}